add precommit/prevote blockState getters, add Set/GetLatestRound to grandpaState
