+++ dot/state/grandpa.go
@@ -24,6 +24,7 @@ import (
 
 	"github.com/ChainSafe/chaindb"
 	"github.com/ChainSafe/gossamer/dot/types"
+	"github.com/ChainSafe/gossamer/lib/common"
 	"github.com/ChainSafe/gossamer/lib/scale"
 )
 
@@ -49,18 +50,19 @@ func NewGrandpaStateFromGenesis(db chaindb.Database, genesisAuthorities []*types
 		db: grandpaDB,
 	}
 
-	err := s.setCurrentSetID(genesisSetID)
-	if err != nil {
+	if err := s.setCurrentSetID(genesisSetID); err != nil {
 		return nil, err
 	}
 
-	err = s.setAuthorities(genesisSetID, genesisAuthorities)
-	if err != nil {
+	if err := s.SetLatestRound(0); err != nil {
 		return nil, err
 	}
 
-	err = s.setSetIDChangeAtBlock(genesisSetID, big.NewInt(0))
-	if err != nil {
+	if err := s.setAuthorities(genesisSetID, genesisAuthorities); err != nil {
+		return nil, err
+	}
+
+	if err := s.setSetIDChangeAtBlock(genesisSetID, big.NewInt(0)); err != nil {
 		return nil, err
 	}
 
@@ -136,6 +138,24 @@ func (s *GrandpaState) GetCurrentSetID() (uint64, error) {
 	}
 
 	return binary.LittleEndian.Uint64(id), nil
+}
+
+// SetLatestRound sets the latest finalised GRANDPA round in the db
+func (s *GrandpaState) SetLatestRound(round uint64) error {
+	buf := make([]byte, 8)
+	binary.LittleEndian.PutUint64(buf, round)
+	return s.db.Put(common.LatestFinalizedRoundKey, buf)
+}
+
+// GetLatestRound gets the latest finalised GRANDPA roundfrom the db
+func (s *GrandpaState) GetLatestRound() (uint64, error) {
+	r, err := s.db.Get(common.LatestFinalizedRoundKey)
+	if err != nil {
+		return 0, err
+	}
+
+	round := binary.LittleEndian.Uint64(r[:8])
+	return round, nil
 }
 
 // SetNextChange sets the next authority change
